+++ ReleaseNotes.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## 2.0.1
+## 2.1.0
 
 - Bug fix: Fixed a bug when creating a tenant that had tenant roles
 - BREAKING CHANGE: GetAllRoleNamesAsync is now called GetRoleNamesForUsersAsync and takes the UserId